services/backlog-service/src/requirements.txt
fastapi
uvicorn
psycopg2-binary
structlog
httpx
redis
tenacity
